Added UI

+++ Source/EnemyBehavior.h
@@ -29,7 +29,7 @@ public:
 	{
 		Vec2F Distance = Target->Position - TheEnemy->Position;
 		if (Distance.getLength() < 800.0) {
-			TheEnemy->TargetDirection = Approach<Angle, double>(TheEnemy->TargetDirection, Distance.getAngle(), t / 2.0f);		
+			TheEnemy->TargetDirection = Approach<Angle, float>(TheEnemy->TargetDirection, Distance.getAngle(), t / 2.0f);		
 			TheEnemy->Acceleration = TheEnemy->TargetDirection.toDirection() * Speed;
 		}		
 	}
